Revert layouts to old style
